Split store implementation
